Add support for subroutines operating in multiple scopes
The current way we set annotations only allows for subroutines to be used
in a single vcl state function.

Fastly is allows that and it also leads to copy-pasting functions with different
annotations, which is not a good pattern.

I only implemented them as annotations in the form `@<state_name>` instead of supporting them in the
function name because I think it will be weird to have `sub foo_recv_miss_deliver`

Signed-off-by: Sotiris Nanopoulos <sotiris.nanopoulos@reddit.com>