Fix testing.override_host
Currently, trying to use `testing.override_host()` in a test results in
the following error:
```
$ falco test -I . default.vcl
Running tests... Done.
 FAIL  /Users/mmalone/src/go/src/github.com/ysugimoto/falco/examples/testing/default.test.vcl
  ●  [RECV]  Foo request header should contains "hoge"

    [RuntimeException] Function testing.override_host is not defined in/Users/mmalone/src/go/src/github.com/ysugimoto/falco/examples/testing/default.test.vcl at line: 5, position: 3
```

I tried writing a unit test for the `tester.Run()` method, but it's
tricky due to the global variables `injectedVariable` and
`builtinFunctions`. Making those part of the `Interpreter` interface
would fix that.