Fix backend is not found in director
This should fix https://github.com/ysugimoto/falco/issues/196, or at
least one of the causes of these errors. The
problem is that `ProcessDeclarations()` method parses the declarations in order,
and if a `director` declaration references a `backend` declaration
that's defined later, it'll trigger `backend 'F_Custom_backend' is
not found`. The fix is to have `ProcessDeclarations()` do two passes:
the first to process backends and the second for everything else.