Fix range for randomint() and randomint_seeded()
The [rand.Int63n(n)](https://pkg.go.dev/math/rand#Rand.Int63n) function
returns an integer in the range `[0, n)`, but
[randomint(from, to)](https://developer.fastly.com/reference/vcl/functions/randomness/randomint/)
returns an integer in the range `[from, to]`. Therefore, we need to add
one to `n` to get the full range.

Also, I fixed a call to the deprecated
[rand.Seed()](https://pkg.go.dev/math/rand#Seed) function.