Store function state in a stack
As functions are called store the current execution state for the
function in a stack. This includes the local variables and re.group.{N}
variables.

This stack is used to detect recursion and prevent an interpreter panic.

In the future this information could be used for including a VCL stack
trace in error output or for accessing calling function state in the
debugger.