Handle implicit type conversion when calling functions
This fixes errors like `[regsub] Argument 1 expects STRING type but
BACKEND provided` when calling builtin functions and passing a
non-string for a string argument. Fastly handles that by implicitly converting the
argument to a string:
> These types all have implicit conversions to strings, such that their values may be used in contexts where a STRING value is necessary.
https://www.fastly.com/documentation/reference/vcl/types/

I wasn't sure the cleanist way of implementing. Originally, I was
thinking of doing this conversion in `ProcessFunctionCallExpression()` and
`ProcessFunctionCallStatement()` (the two places that call functions)
like we're already doing for the `IsIdentArgument()` check, but it
needs to know the argument types for every function to do that, which
means encoding that information in the `Function` struct.