Lex and parse long strings as multiple tokens
The previous iteration of the long string parsing improvement to handle
heredoc-delimited strings (e.g. {foo"bar"foo}) emitted a single `STRING`
from the lexer. This loses the delimiter value, which we need to be able
to round trip and emit the semantically-same VCL from the formatter.

This work adds new token types for the opening and closing long string
delimiters, and modifies ast.String to carry delimiter information.